requirements.txt
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
beautifulsoup4==4.12.2
python-frontmatter==1.0.0
PyYAML==6.0.1
fastapi==0.100.0
uvicorn==0.22.0
mercadopago==2.2.0
python-dotenv==1.0.0
